multiagent-dev-platform/apps/orchestrator_api/requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# Multiagent Dev Platform - Python Dependencies
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

# ━━━ Web Framework ━━━
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ━━━ Database ━━━
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ━━━ Vector Database (Qdrant) ━━━
qdrant-client==1.7.3

# ━━━ LLM Integration ━━━
openai==1.10.0  # Compatible with Ollama API
anthropic==0.8.1
google-generativeai==0.3.2
httpx==0.26.0

# ━━━ Embeddings & NLP ━━━
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.2+cpu; platform_system=="Linux"
sentence-transformers==2.3.1
tiktoken==0.5.2

# ━━━ Git Operations ━━━
gitpython==3.1.41

# ━━━ File Operations ━━━
watchdog==4.0.0
pathspec==0.12.1

# ━━━ YAML/Config ━━━
pyyaml==6.0.1
python-dotenv==1.0.0

# ━━━ Utilities ━━━
python-multipart==0.0.6
aiofiles==23.2.1
jinja2==3.1.3
markdown==3.5.2

# ━━━ Testing & Validation ━━━
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# ━━━ Logging & Monitoring ━━━
loguru==0.7.2
python-json-logger==2.0.7

# ━━━ Security ━━━
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ━━━ Async ━━━
asyncio==3.4.3
aiohttp==3.9.1

# ━━━ Code Analysis ━━━
radon==6.0.1
pylint==3.0.3
